requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
aiohttp-socks==0.5.3
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
cchardet==2.1.6
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
elasticsearch==7.8.1
fake-useragent==0.1.11
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.1.2
geographiclib==1.50
geopy==2.0.0
googletransx==2.4.2
idna==2.10
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.11.2
joblib==0.16.0
lxml==4.5.2
MarkupSafe==1.1.1
multidict==4.7.6
newspaper3k==0.2.8
nltk==3.5
numpy==1.19.1
pandas==1.1.0
Pillow==7.2.0
pycares==3.1.1
pycparser==2.20
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-file==1.5.1
schedule==0.6.0
six==1.15.0
soupsieve==2.0.1
tinysegmenter==0.3
tldextract==2.2.3
tqdm==4.48.2
twint==2.1.20
typing-extensions==3.7.4.2
urllib3==1.25.10
Werkzeug==1.0.1
yarl==1.5.1
